It was an f32 overflow error hahahahahaha ... srs

+++ poker/src/equity.rs
@@ -2,9 +2,9 @@
 // Results of player i at index i.
 #[derive(Debug, Clone)]
 pub struct EquityResults {
-    pub wins:   Vec<f32>,
-    pub ties:   Vec<f32>,
-    pub total:  f32,
+    pub wins:   Vec<f64>,
+    pub ties:   Vec<f64>,
+    pub total:  f64,
 }
 
 impl EquityResults {
@@ -17,7 +17,7 @@ impl EquityResults {
         }
     }
 
-    pub fn equities(&self) -> Vec<f32> {
+    pub fn equities(&self) -> Vec<f64> {
         let mut equities = vec![0.0; self.wins.len()];
         for i in 0..self.wins.len() {
             equities[i] = ((self.wins[i] + self.ties[i] / 2.0) / self.total) * 100.0;
@@ -25,7 +25,7 @@ impl EquityResults {
         equities
     }
 
-    pub fn win_pct(&self) -> Vec<f32> {
+    pub fn win_pct(&self) -> Vec<f64> {
         let mut win_pct = vec![0.0; self.wins.len()];
         for i in 0..self.wins.len() {
             win_pct[i] = self.wins[i] / self.total * 100.0;
@@ -33,7 +33,7 @@ impl EquityResults {
         win_pct
     }
 
-    pub fn tie_pct(&self) -> Vec<f32> {
+    pub fn tie_pct(&self) -> Vec<f64> {
         let mut tie_pct = vec![0.0; self.wins.len()];
         for i in 0..self.wins.len() {
             tie_pct[i] = self.ties[i] / self.total * 100.0;
